tugas 6 done

+++ todolist/views.py
@@ -1,11 +1,12 @@
 import datetime
-from django.http import HttpResponseRedirect
+from django.http import HttpResponse, HttpResponseRedirect, JsonResponse
 from django.urls import reverse
 from django.shortcuts import render, redirect
 from django.contrib.auth.forms import UserCreationForm
 from django.contrib import messages
 from django.contrib.auth import authenticate, login, logout
 from django.contrib.auth.decorators import login_required
+from django.core import serializers
 from todolist.models import Task
 
 # Create your views here.
@@ -69,17 +70,75 @@ def create_task(request):
 
 @login_required(login_url="/todolist/login/")
 def delete_task(request, id):
-    task = Task.objects.get(user=request.user, id=id)
-    task.delete()
-    return HttpResponseRedirect(reverse("todolist:show_todolist"))
+    if request.method == "DELETE":
+        task = Task.objects.get(user=request.user, id=id)
+        task.delete()
+        return JsonResponse(
+            {
+                "pk": task.id,
+                "fields": {
+                    "title": task.title,
+                    "description": task.description,
+                    "is_finished": task.is_finished,
+                    "date": task.date,
+                },
+            },
+            status=200,
+        )
+
+    #task = Task.objects.get(user=request.user, id=id)
+    #task.delete()
+    #return HttpResponseRedirect(reverse("todolist:show_todolist"))
 
 @login_required(login_url="/todolist/login/")
 def finish_task(request, id):
-    task = Task.objects.get(user=request.user, id=id)
-    task.is_finished = not task.is_finished
-    task.save(update_fields=["is_finished"])
-    return HttpResponseRedirect(reverse("todolist:show_todolist"))
+    if request.method == "PUT":
+        task = Task.objects.get(user=request.user, id=id)
+        task.is_finished = not task.is_finished
+        task.save()
+        return JsonResponse(
+            {
+                "pk": task.id,
+                "fields": {
+                    "title": task.title,
+                    "description": task.description,
+                    "is_finished": task.is_finished,
+                    "date": task.date,
+                },
+            },
+            status=200,
+        )
+    # task.save(update_fields=["is_finished"])
+    # return HttpResponseRedirect(reverse("todolist:show_todolist"))
 
 def todolist(request):
     randomclass = random.choice(['card border-primary mb-3', 'card border-secondary mb-3', 'card border-success mb-3', 'card border-danger mb-3'])
     return render(request, "todolist.html", {'randomclass': randomclass})
+
+@login_required(login_url="/todolist/login/")
+def show_json(request):
+    task = Task.objects.filter(user=request.user)
+    return HttpResponse(serializers.serialize("json", task), content_type="application/json")
+
+def add_task(request):
+    if request.method == "POST":
+        title = request.POST.get("title")
+        description = request.POST.get("description")
+        task = Task.objects.create(
+            user=request.user,
+            title=title,
+            description=description,
+            date=datetime.datetime.today(),
+        )
+        return JsonResponse(
+            {
+                "pk": task.id,
+                "fields": {
+                    "title": task.title,
+                    "description": task.description,
+                    "is_finished": task.is_finished,
+                    "date": task.date,
+                },
+            },
+            status=200
+        )
